requirements.txt
SQLAlchemy>=1.4,<3
alembic>=1.10
psycopg2-binary>=2.9
python-dotenv>=1.0
